Rollback to the first stable version.
I forked this repository from Rexarrior's account
and reset to the version I made (this is written almost
solely be me, excepting 2 or 3 lines of code). This
is the first stable version + Rexarrior's bugfix in
Connection.cs.

Repository is now finally properly named -
Universal Connector - this is a meaningful name for
the project, contrary to the old LVNS (actually, it
was named so, because its original purposes were
quite different (then they became totally unclear)).
If you care about that old name, check
https://github.com/Rexarrior/LVNS older versions.
So, .pdf and .vsdx files in the root were removed
and README.md modified.

Old "Core" directory removed and "SimplifiedCore"
renamed to just "Core".

Directory "InterfacesMonitors" created to store DLLs
and probably their sources - the DLLs that interact
with certain interfaces and are used by Core.

ConstDefs.cs improved - "Numerics" class added. Now
we store buffer sizes and timeouts there. If anything
falls into 'numeric constant'" category, it should be
declared here.

Also, in transfer loop (Connection.cs) a delay
(Thread.Sleep) added - that was an advice from
Rexarrior to reduce CPU usage. I think in Release
versions this line (this delay is only a single line
of code) should be commented out, to maximise the
speed of data transfering.

In .gitignore there now is Solutions folder -
this is the expected name of the folder where
some files needed in a particular application
or solution will be stored.

Thank you for reading this commit message
all the way through. I appreciate your time.

=====
Believe in yourself,

KKastaneda, 05 December, 2018